+++ src/cpp/session/modules/SessionFiles.cpp
@@ -23,6 +23,7 @@
 #include <fstream>
 #include <sstream>
 #include <algorithm>
+#include <gsl/gsl>
 
 #include <boost/lexical_cast.hpp>
 
@@ -482,7 +483,7 @@ void handleFilesRequest(const http::Request& request,
    }
    
    // compute path to file
-   int prefixLen = prefix.length();
+   int prefixLen = gsl::narrow_cast<int>(prefix.length());
    std::string relativePath = http::util::urlDecode(uri.substr(prefixLen));
    if (relativePath.empty())
    {
